Move file format definitions into shared module
