+++ app/src/main/java/kolskypavel/ardfmanager/backend/files/constants/DataFormat.kt
@@ -1,13 +1,3 @@
 package kolskypavel.ardfmanager.backend.files.constants
 
-enum class DataFormat(var value: Int) {
-    TXT(0),
-    HTML(1),
-    JSON(2),
-    CSV(3),
-    IOF_XML(4);
-
-    companion object {
-        fun getByValue(value: Int) = DataFormat.entries.firstOrNull { it.value == value }
-    }
-}
+typealias DataFormat = org.openardf.radioomanager.shared.files.DataFormat
