+++ app/src/main/java/kolskypavel/ardfmanager/backend/files/constants/DataType.kt
@@ -1,14 +1,3 @@
 package kolskypavel.ardfmanager.backend.files.constants
 
-enum class DataType(var value: Int) {
-    CATEGORIES(0),
-    COMPETITORS(1),
-    COMPETITOR_STARTS(2),
-    RESULTS_LIVE(3),
-    READOUT_DATA(4),
-    RESULTS_FINAL (5);
-
-    companion object {
-        fun getByValue(value: Int) = DataType.entries.firstOrNull { it.value == value }
-    }
-}
+typealias DataType = org.openardf.radioomanager.shared.files.DataType
