+++ shared/src/commonMain/kotlin/org/openardf/radioomanager/shared/files/DataFormat.kt
@@ -1,0 +1,13 @@
+package org.openardf.radioomanager.shared.files
+
+enum class DataFormat(var value: Int) {
+    TXT(0),
+    HTML(1),
+    JSON(2),
+    CSV(3),
+    IOF_XML(4);
+
+    companion object {
+        fun getByValue(value: Int) = entries.firstOrNull { it.value == value }
+    }
+}
